+++ src/vertical-scenes-dock.hpp
@@ -5,7 +5,6 @@
 class QListWidget;
 class ShortsDock;
 
-/* Lists main OBS scenes (not a parallel vertical scene collection). */
 class VerticalScenesDock : public QFrame {
 	Q_OBJECT
 
@@ -15,6 +14,10 @@ public:
 private slots:
 	void RefreshList();
 	void OnSelectionChanged();
+	void OnAdd();
+	void OnRemove();
+	void OnDuplicate();
+	void OnRename();
 
 private:
 	void BuildUI();